==> 5_projekt_koncowy/rozwiazania/2_time_series_analysis.ipynb <==
# Time series analysis

Celem tej części projektu, jest stworzenie modelu szeregu czasowego dla całkowitej liczby wypożyczeń.

Zadania:
1. Przeprowadź wstępną analizę danych w tym:
    - pobranie danych i podstawowe statystyki
    - reasmpling na dane dzienne
    - dekompozycję
    - zagreguj dane do miesięcznych, dla modelu długookresowego.
2. Stwórz model długookresowy:
    - podziel dane na zbiór treningowy oraz testowy
    - stwórz model miesięcznej predykcji wypożyczeń w kolejnym roku 
    - dokonaj predykcji na danych historycznych oraz 2021 rok (przedstaw graficznie wyniki)
    - oceń jakość za pomocą dwóch, wybranych wskaźników dla modeli regresyjnych.
3. Model krótkookresowy - predykcja na kolejny dzień:
Drugi model będzie potrzebny, aby na koniec danego dnia mieć oszacowanie jak bardzo system będzie oblegany. Można dzięki temu zaplanować serwis stacji, rowerów, czy przemieścić rowery tak, aby najbardziej oblegane stacje miały dostatecznie dużo pojazdów.
    - Dodaj do danych zmienne egzogeniczne (średnie wartości z poprzedniego dnia, temperatura z ostatnich 7 dni)
    - podziel dane na zbiór treningowy oraz testowy
    - Stwórz model dedykowany do szeregów czasowych ze zmiennymi egzogenicznymi.
    - Oceń model.
    - Porównaj model do podejścia machine learning.

## 1. Przeprowadź wstępną analizę danych.

- pobranie danych i podstawowe statystyki

import pandas as pd 
import statsmodels.api as sm
import matplotlib.pyplot as plt

# puść ten kod, 
# jeżeli wywołujesz plik  w folderze rozwiąznaia, 
# a ramka danych znajduje się w folderze data
import os 
os.chdir('../')

# Pobranie danych
df = pd.read_parquet('data/total_agg.parquet')

# head
df.head()

# info
df.info()

- reasmpling na dane dzienne


# ustawienie daty jako indeks
df = df.set_index('departure_date')

# resampling
df_resampled = df.resample('D').sum().fillna(0)
df_resampled

- dekompozycja

# obiekt dekompozycji
decomp = sm.tsa.seasonal_decompose(df_resampled['numbers_of_renting'],model='additive',period= 365)

# wykres dekompozycji
decomp.plot()
plt.show()

- zagreguj dane do miesięcznych, dla modelu długookresowego.

df_monthly = df.resample('ME').sum().fillna(0).reset_index().rename(columns = {'numbers_of_renting': 'y',
                                                                               'departure_date':'ds'})
df_monthly

# dekompozycja miesięczna
decom_monthly = sm.tsa.seasonal_decompose(df_monthly['y'],period=12)
decom_monthly.plot()
plt.show()

2. Stwórz model długookresowy:
    - podziel dane na zbiór treningowy oraz testowy

# train / test split
train_monthly = df_monthly.loc[df_monthly['ds'].dt.year<2020,['y','ds']]
test_monthly = df_monthly.loc[df_monthly['ds'].dt.year==2020,['y','ds']]

- stwórz model miesięcznej predykcji wypożyczeń w kolejnym roku 

# Pobranie neuralprophet
from neuralprophet import NeuralProphet
import numpy as np

# obiekt modelu
model_monthly = NeuralProphet()

# fit
model_monthly.fit(train_monthly, freq = 'ME')

- dokonaj predykcji na danych historycznych oraz 2021 rok (przedstaw graficznie wyniki)

# forecast
forecast_historical_monthly = model_monthly.predict(df_monthly[['ds','y']])
forecast_historical_monthly

# Stworzenie zakresu danych dla roku 2021
new_data_index = pd.date_range(start = df_monthly['ds'].max()+ pd.Timedelta(days=30),end = df_monthly['ds'].max()+pd.Timedelta(days=400),freq= 'ME')
new_data_index

# Stworzenie ramki danych
forecast_df_monthly = pd.DataFrame(data=new_data_index, columns = ['ds'])
forecast_df_monthly['y'] = np.nan
forecast_df_monthly

# Predykcja przyszłego roku
forecast_future_monthly = model_monthly.predict(forecast_df_monthly)

forecast_future_monthly

# funkcja do rysowania wykresu
def plot_forecasts(df: pd.DataFrame, 
                   date_col :str= 'ds',
                   actuals_col: str = 'y',
                   pred_col: str='yhat1'):
    plt.figure(figsize=(10,6))
    plt.plot(df[date_col],df[actuals_col],label = 'actual')
    plt.plot(df[date_col],df[pred_col],label = 'prediction')
    plt.legend()
    plt.show()

# Wykres
plot_forecasts(pd.concat([forecast_historical_monthly,forecast_future_monthly]))

# Dodanie miesiąca
forecast_historical_monthly['month'] = forecast_historical_monthly.ds.dt.month
forecast_future_monthly['month'] = forecast_future_monthly.ds.dt.month

# Wyzerowanie forecastów w miesiącach, w których system nie funkcjonuje
forecast_historical_monthly.loc[forecast_historical_monthly['month'].isin([12,1,2]),'yhat1'] = 0
forecast_future_monthly.loc[forecast_future_monthly['month'].isin([12,1,2]),'yhat1'] = 0

# wykres
plot_forecasts(pd.concat([forecast_historical_monthly, forecast_future_monthly]))

# dodanie znacznika zbioru train/ test
forecast_historical_monthly['set'] = 'train'
forecast_historical_monthly.loc[forecast_historical_monthly['ds']>='2020-01-01','set'] = 'test'

# Pobranie metryk
from sklearn.metrics import root_mean_squared_error, r2_score

# r2 score
forecast_historical_monthly.loc[:,['y','yhat1','set']].groupby('set').apply(lambda x: r2_score(x['y'],x['yhat1']))

# rmse
forecast_historical_monthly.loc[:,['y','yhat1','set']].groupby('set').apply(lambda x: root_mean_squared_error(x['y'],x['yhat1']))

forecast_historical_monthly.describe()

import os 
import joblib

if not os.path.exists('models'):
    os.mkdir('models')

joblib.dump(model_monthly, 'models/model_monthly_forecasts.joblib')

## 3. Model krótkookresowy - predykcja na kolejny dzień:
Drugi model będzie potrzebny, aby na koniec danego dnia mieć oszacowanie jak bardzo system będzie oblegany. Można dzięki temu zaplanować serwis stacji, rowerów, czy przemieścić rowery tak, aby najbardziej oblegane stacje miały dostatecznie dużo pojazdów.

- Dodaj do danych zmienne egzogeniczne (średnie wartości z poprzedniego dnia, temperatura z ostatnich 7 dni)

df_resampled = df_resampled.sort_index()

df_resampled.head()

# dodanie zmienych z wczoraj
df_resampled['temperature_yesterday'] = df_resampled['Air temperature (degC)'].shift(1)
df_resampled['avg_speed_yesterday'] = df_resampled['avg_speed (km/h)'].shift(1)
df_resampled['avg_duration_yesterday'] = df_resampled['duration (sec.)'].shift(1)

# dodanie informacji z daty
df_resampled['month'] = df_resampled.index.month
df_resampled['day'] = df_resampled.index.day
df_resampled['quarter'] = df_resampled.index.quarter

# usunięcie braków danych
df_resampled = df_resampled.dropna().reset_index()
df_resampled

# dopasowanie nazw w ramce danych
df_resampled = df_resampled.rename(columns= {'departure_date':'ds','numbers_of_renting':'y'})

df_resampled

- podziel dane na zbiór treningowy oraz testowy


# train / test split
train_daily = df_resampled[df_resampled['ds'].dt.year < 2020]
test_daily = df_resampled[df_resampled['ds'].dt.year==2020]

# zmienne do modelowania
cols_names =  ['ds','y','avg_speed_yesterday','avg_duration_yesterday','Air temperature (degC)']

train_daily[cols_names]

- Stwórz model dedykowany do szeregów czasowych ze zmiennymi egzogenicznymi.


# model na danych dziennych 
model_daily = NeuralProphet()


# dodanie regresorów
model_daily.add_future_regressor('avg_speed_yesterday')
model_daily.add_future_regressor('avg_duration_yesterday')
model_daily.add_lagged_regressor('Air temperature (degC)',n_lags=7)
model_daily.add_country_holidays('Finland')

# fit
model_daily.fit(train_daily[cols_names])

- Oceń model.


# predykcje historyczne
historical_predictions = model_daily.predict(df_resampled[cols_names])

# dodanie miesiąca oraz identyfikatora train/test
historical_predictions['month'] = historical_predictions.ds.dt.month
historical_predictions['set'] = 'train'
historical_predictions.loc[historical_predictions['ds']>='2020-01-01','set'] = 'test'

# wykres
plot_forecasts(historical_predictions)

# wyzerowanie miesięcy z zamkniętym systemem
historical_predictions.loc[historical_predictions['month'].isin([12,1,2]),'yhat1'] = 0

# ponowny wykres
plot_forecasts(historical_predictions)

historical_predictions[['yhat1']].isna().max()

historical_predictions[historical_predictions['yhat1'].isna()]

# usunięcie braków danych
historical_predictions  = historical_predictions.dropna()

historical_predictions.shape

# r2 score
historical_predictions.loc[:,['y','yhat1','set']].groupby('set').apply(lambda x:r2_score(x['y'],x['yhat1']))

# rmse
historical_predictions.loc[:,['y','yhat1','set']].groupby('set').apply(lambda x:root_mean_squared_error(x['y'],x['yhat1']))

historical_predictions.describe()

- Porównaj model do podejścia machine learning.

df_resampled = df.resample('D').sum().fillna(0)

df_resampled.columns

# dodanie zmienych z wczoraj
df_resampled = df_resampled.sort_index()
df_resampled['temperature_yesterday'] = df_resampled['Air temperature (degC)'].shift(1)
df_resampled['avg_speed_yesterday'] = df_resampled['avg_speed (km/h)'].shift(1)
df_resampled['avg_duration_yesterday'] = df_resampled['duration (sec.)'].shift(1)
df_resampled = df_resampled.rename(columns = {'departure_date':'ds', 'numbers_of_renting':'y' })
df_resampled['y_yesterday'] = df_resampled['y'].shift(1)
# dodanie informacji z daty
df_resampled['month'] = df_resampled.index.month
df_resampled['day'] = df_resampled.index.day
df_resampled['quarter'] = df_resampled.index.quarter
# usunięcie braków danych
df_resampled = df_resampled.dropna().reset_index()
df_resampled = df_resampled.rename(columns = {'departure_date':'ds'})
df_resampled

# Filtrowanie tylko niepustych wartości
df_resampled_non_zero= df_resampled[df_resampled['y']>0].reset_index(drop= True)

# train/test split
train_daily  =df_resampled_non_zero[df_resampled_non_zero['ds'].dt.year<2020]
test_daily = df_resampled_non_zero[df_resampled_non_zero['ds'].dt.year ==2020]

train_daily.columns

# zmienne x
x_names = ['temperature_yesterday',
       'avg_speed_yesterday', 'avg_duration_yesterday','month', 'day',
       'quarter', 'y_yesterday']

# train_x / train_y
train_x = train_daily[x_names]
train_y = train_daily['y']
test_x  = test_daily[x_names]
test_y = test_daily['y']

# pobranie funkcji
from sklearn.ensemble import RandomForestRegressor
from scipy.optimize import differential_evolution

# Granice przeszukiwan
bounds = [(50,200),
          (5,50),
          (5,50)]

# Funkcja optymalizacyjna
def optimization_function(params, train_x, train_y, test_x,test_y):
    params = {'n_estimators': int(round(params[0])),
    
              'max_depth': int(round(params[1])),
              'min_samples_split': int(round(params[2]))
              }
    model = RandomForestRegressor(**params).fit(train_x,train_y)
    preds = model.predict(test_x)
    return -r2_score(test_y,preds)

# Optymalizacja ewolucyjna
optimization = differential_evolution(func = optimization_function,
                                      bounds = bounds,
                                      args = (train_x, train_y,test_x,test_y),
                                      maxiter=10)

# Najlepsze parametry
params = optimization.x
best_params ={'n_estimators': int(round(params[0])),
              'max_depth': int(round(params[1])),
              'min_samples_split': int(round(params[2]))}

best_params

# finalny model
model_final = RandomForestRegressor(**best_params).fit(train_x,train_y)

# predykcje
df_resampled_non_zero['pred_RF'] = model_final.predict(df_resampled_non_zero[x_names])

# wykres 
plt.figure(figsize=(10,6))
plt.plot(historical_predictions['ds'],historical_predictions['y'],label=  'actual')
plt.plot(df_resampled_non_zero['ds'], df_resampled_non_zero['pred_RF'],label = 'prediction_RF')
plt.legend()
plt.show()

# wykres
plt.figure(figsize=(10,6))
plt.plot(historical_predictions['ds'],historical_predictions['y'],label=  'actual')
plt.plot(df_resampled_non_zero['ds'], df_resampled_non_zero['pred_RF'],label = 'prediction_RF')
plt.plot(historical_predictions['ds'],historical_predictions['yhat1'],label='prediction_NP', alpha=0.6)
plt.legend()
plt.show()

# znacznik train/test
df_resampled_non_zero['set'] = 'train'
df_resampled_non_zero.loc[df_resampled_non_zero['ds']>='2020-01-01','set'] = 'test'

# r2_score
df_resampled_non_zero.loc[:,['y','pred_RF','set']].groupby('set').apply(lambda x: r2_score(x['y'],x['pred_RF']))

# r2 score na neural prophet
historical_predictions.loc[historical_predictions['y']>0,['y','yhat1','set']].groupby('set').apply(lambda x: r2_score(x['y'],x['yhat1']))

# biblioteki do zapisania modeli
import joblib
import os 

# stworzenie ścieżki
if not os.path.exists('models'):
    os.mkdir('models')

# zapisanie modelu krótkookresowego
joblib.dump(model_final, 'models/model_RF_daily_forecasts.joblib')


# zapisanie modelu długookresowego
