==> 5_projekt_koncowy/rozwiazania/4_prognozowanie.ipynb <==
# Prognozowanie

Przygotowanie funkcji do prognozowania wykonamy dla modelu klasyfikacyjnego oraz krótkookresowego.

Zadania do wykonania:
1. Zapis danych za ostatnie dni września 2020 do lokalnej bazy postgres.
2. Stworzenie funkcji do pobierania modelu. (funkcja load_model z help_function.py)
3. Stworzenie funkcji do pobierania i agregacji danych dla prognozy na następną godzinę (klasyfikacja). (funkcje: data_load, agg_df_to_hour_station z help_function.py)
4. Stworzenie funkcji do pobierania i agregacji danych dla prognozy na kolejny dzień. (funkcje: agg_df_to_daily, prepare_data_daily z help_function.py)
5. Stworzenie całościowej funkcji, która: 
    - pobierze od użytkownika informacje na jaki termin potrzebna jest predykcja.
    - pobierze dane oraz model.
    - przygotuje dane do predykcji.
    - dokona predykcji i zapisze je na dysku (dla modelu klasyfikacyjnego).
    - wyświetli predykcje dla wybranej przez użytkownika stacji.
    - wyświetli predykcje na kolejny dzień.
    - Funkcje forecast_process, forecast_hourly_station, forecast_daily z aplikacja.py

1. Zapis danych za ostatnie dni września 2020 do lokalnej bazy postgres.

# import bibliotek
import pandas as pd 
from sqlalchemy import create_engine 
import os

# puść ten kod, 
# jeżeli wywołujesz plik  w folderze rozwiąznaia, 
# a ramka danych znajduje się w folderze data
import os 
os.chdir('../')

# skopiować do pliku .env
#DB = 'postgresql://postgres:postgres@localhost'

# zaladowanie zmiennych srodowiskowych
from dotenv import load_dotenv 
load_dotenv()

# stworzenie engine
DB = os.getenv("DB")
DB


engine = create_engine(DB)

# pobranie danych
df = pd.read_parquet('data/source_file.parquet')

# info
df.info()

# przefiltrowanie danych
df_to_sql = df[(df['departure_date']>='2020-09-22') & (df['departure_date']<='2020-09-23')]
df_to_sql.shape

# załadowanie danych
df_to_sql.to_sql('renting_data', con= engine, method = 'multi', chunksize=10000, if_exists='replace')


# pobranie w celu sprawdzenia
df_check = pd.read_sql("select * from renting_data",con = engine)

# shape
df_check.shape

df_check.head()

2. Stworzenie funkcji do pobierania modelu. (funkcja load_model z help_function.py)


# pobranie funkcji
from help_function import load_model

# sprawdzenie pobierania
model = load_model()

model

load_model('model')

3. Stworzenie funkcji do pobierania i agregacji danych dla prognozy na następną godzinę (klasyfikacja). (funkcje: data_load, agg_df_to_hour_station z help_function.py)


# pobranie funkcji
from help_function import data_load, agg_df_to_hour_station

df = data_load("departure_date='2020-09-23'")
df.head()

df_new = agg_df_to_hour_station(df)

df_new.head()

4. Stworzenie funkcji do pobierania i agregacji danych dla prognozy na kolejny dzień. (funkcje: agg_df_to_daily, prepare_data_daily z help_function.py)


# pobranie funkcji
from help_function import agg_df_to_daily, prepare_data_daily, data_load

# wywołanie
df = data_load("departure_date>='2020-09-01'")
df.head()

import os
os.getcwd()

df_daily  = agg_df_to_daily(df)
df_daily

df_daily_2 = prepare_data_daily(df_daily)

df_daily_2

5. Stworzenie całościowej funkcji, która: 
    - pobierze od użytkownika informacje na jaki termin potrzebna jest predykcja.
    - pobierze dane oraz model.
    - przygotuje dane do predykcji.
    - dokona predykcji i zapisze je na dysku (dla modelu klasyfikacyjnego).
    - wyświetli predykcje dla wybranej przez użytkownika stacji.
    - wyświetli predykcje na kolejny dzień.
    - Funkcje forecast_process, forecast_hourly_station, forecast_daily z aplikacja.py

# pobranie funkcji
from aplikacja import forecast_process

# wywolanie
forecast_process()
